// src/shared/api/generated/model/routePoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */
import type { OperationType } from './operationType'
import type { RoutePointCargo } from './routePointCargo'
import type { RoutePointContact } from './routePointContact'
import type { RoutePointLocation } from './routePointLocation'

export interface RoutePoint {
  row_num?: number
  op_type?: OperationType
  start_date?: string
  end_date?: string
  /** @nullable */
  comment?: string | null
  contractor?: string
  contractor_inn?: string
  location?: RoutePointLocation
  cargo?: RoutePointCargo
  contact?: RoutePointContact
}